backend/requirements.txt
fastapi==0.115.11
uvicorn[standard]==0.34.0
pydantic==2.10.6
httpx==0.28.1
python-dotenv==1.2.2
pytest==8.3.5
